+++ src/lib.rs
@@ -1,6 +1,5 @@
 extern crate rustc_serialize;
 extern crate docopt;
-extern crate exec;
 extern crate regex;
 
 #[cfg(test)]